server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
pydantic==1.10.14
python-dotenv==1.0.1

faiss-cpu
farm-haystack==1.26.2
sentence-transformers==3.0.1
# transformers==4.44.2
torch>=2.1.0

# Database dependencies
SQLAlchemy<2.0

pypdf
python-docx

openai>=1.0.0

rank-bm25

# Optional: enable BM25 later (comment out if you don't want Java/Lucene deps)
# pyserini==0.40.0
